Fallback to file-based env collector on Windows
