+++ internal/command/command_unix_test.go
@@ -19,8 +19,6 @@ import (
 )
 
 func init() {
-	EnvDumpCommand = "env -0"
-
 	// Set to false to disable sending signals to process groups in tests.
 	// This can be turned on if setSysProcAttrPgid() is called in Start().
 	SignalToProcessGroup = false
