+++ internal/command/shell_env_collector_windows.go
@@ -1,0 +1,7 @@
+package command
+
+import "io"
+
+func buildShellEnvCollector(w io.Writer) (shellEnvCollector, error) {
+	return newFileShellEnvCollector(w)
+}
